+++ crates/cli/src/commands/pin.rs
@@ -2,30 +2,27 @@ use std::env;
 use std::path::PathBuf;
 
 use clap::Args;
-use proto_core::{load_tool, Id, ToolsConfig, UnresolvedVersionSpec};
-use starbase::system;
+use proto_core::{load_tool, Id, Tool, ToolsConfig, UnresolvedVersionSpec};
+use starbase::{system, SystemResult};
 use starbase_styles::color;
 use tracing::{debug, info};
 
 #[derive(Args, Clone, Debug)]
 pub struct PinArgs {
     #[arg(required = true, help = "ID of tool")]
-    id: Id,
+    pub id: Id,
 
     #[arg(required = true, help = "Version or alias of tool")]
-    spec: UnresolvedVersionSpec,
+    pub spec: UnresolvedVersionSpec,
 
     #[arg(
         long,
         help = "Add to the global user config instead of local .prototools"
     )]
-    global: bool,
+    pub global: bool,
 }
 
-#[system]
-pub async fn pin(args: ArgsRef<PinArgs>) -> SystemResult {
-    let mut tool = load_tool(&args.id).await?;
-
+pub fn internal_pin(tool: &mut Tool, args: &PinArgs) -> SystemResult {
     if args.global {
         tool.manifest.default_version = Some(args.spec.clone());
         tool.manifest.save()?;
@@ -34,12 +31,6 @@ pub async fn pin(args: ArgsRef<PinArgs>) -> SystemResult {
             version = args.spec.to_string(),
             manifest = ?tool.manifest.path,
             "Wrote the global version",
-        );
-
-        info!(
-            "Set the global {} version to {}",
-            tool.get_name(),
-            color::hash(args.spec.to_string())
         );
     } else {
         let local_path = env::current_dir().unwrap_or_else(|_| PathBuf::from("."));
@@ -53,11 +44,20 @@ pub async fn pin(args: ArgsRef<PinArgs>) -> SystemResult {
             config = ?config.path,
             "Wrote the local version",
         );
-
-        info!(
-            "Set the local {} version to {}",
-            tool.get_name(),
-            color::hash(args.spec.to_string())
-        );
     }
+
+    Ok(())
+}
+
+#[system]
+pub async fn pin(args: ArgsRef<PinArgs>) -> SystemResult {
+    let mut tool = load_tool(&args.id).await?;
+
+    internal_pin(&mut tool, args)?;
+
+    info!(
+        "Set the {} version to {}",
+        tool.get_name(),
+        color::hash(args.spec.to_string())
+    );
 }