+++ crates/core/tests/user_config_test.rs
@@ -1,4 +1,4 @@
-use proto_core::{UserConfig, USER_CONFIG_NAME};
+use proto_core::{PinType, UserConfig, USER_CONFIG_NAME};
 use starbase_sandbox::create_empty_sandbox;
 use std::collections::BTreeMap;
 use std::env;
@@ -19,6 +19,7 @@ mod user_config {
                 auto_install: false,
                 node_intercept_globals: true,
                 http: HttpOptions::default(),
+                pin_latest: None,
                 plugins: BTreeMap::default(),
                 path: sandbox.path().join(USER_CONFIG_NAME),
             }
@@ -26,7 +27,7 @@ mod user_config {
     }
 
     #[test]
-    fn can_set_booleans() {
+    fn can_set_values() {
         let sandbox = create_empty_sandbox();
         sandbox.create_file(
             "config.toml",
@@ -34,6 +35,7 @@ mod user_config {
 auto-clean = true
 auto-install = true
 node-intercept-globals = false
+pin-latest = "global"
 "#,
         );
 
@@ -46,6 +48,7 @@ node-intercept-globals = false
                 auto_install: true,
                 node_intercept_globals: false,
                 http: HttpOptions::default(),
+                pin_latest: Some(PinType::Global),
                 plugins: BTreeMap::default(),
                 path: sandbox.path().join(USER_CONFIG_NAME),
             }
@@ -70,6 +73,7 @@ node-intercept-globals = false
                 auto_install: true,
                 node_intercept_globals: false,
                 http: HttpOptions::default(),
+                pin_latest: None,
                 plugins: BTreeMap::default(),
                 path: sandbox.path().join(USER_CONFIG_NAME),
             }